services/llamaindex-service/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
llama-index==0.9.10
llama-index-vector-stores-chroma==0.1.4
llama-index-vector-stores-weaviate==0.1.3
chromadb==0.4.18
weaviate-client==3.25.3
